Flatten project structure
